+++ GUI/QtQuick/Utils.h
@@ -4,7 +4,7 @@
 
 namespace gui_utils {
 
-QString satoshiToQString(double balance);
+QString satoshiToQString(int64_t balance);
 QString normalizedSatoshiToQString(double balance);
 
 }
